requirements.txt
rich
transformers==3.0.2 
pytorchcrf==1.2.0 
torchcontrib==0.0.2
pkuseg==0.0.25